bindings/java: Remove restrictions on param positions
This drops the restrictions on closure/destroy/length parameters being
at specific positions in the list of parameters. These may occur
anywhere in the GIR, so we just make two passes instead -- one to
collect the non-referencing parameters, and then to collect the params
that reference other parameters (that is closures, destroys and array
lengths).